requirements.txt
flask
numpy
librosa
imageio
noisereduce
torch
google-generativeai
transformers
tensorflow
vercel-wsgi